requirements.txt
pillow>=2.5.1
cherrypy
flickrapi
lindh-jsonobject
https://github.com/eblade/jsondb/archive/0.1.tar.gz
